+++ forest/cli/src/cli/mod.rs
@@ -14,7 +14,7 @@ mod db_cmd;
 mod fetch_params_cmd;
 mod mpool_cmd;
 mod net_cmd;
-mod send_cmd;
+pub mod send_cmd;
 mod shutdown_cmd;
 mod snapshot_cmd;
 mod state_cmd;