+++ forest/cli/src/cli/snapshot_cmd.rs
@@ -4,6 +4,7 @@
 use std::{fs, path::PathBuf, sync::Arc, time::Duration};
 
 use anyhow::bail;
+use chrono::{Datelike, Utc};
 use clap::Subcommand;
 use dialoguer::{theme::ColorfulTheme, Confirm};
 use forest_blocks::{tipset_keys_json::TipsetKeysJson, Tipset, TipsetKeys};
@@ -13,7 +14,7 @@ use forest_cli_shared::cli::{
 };
 use forest_db::db_engine::{db_root, open_proxy_db};
 use forest_genesis::{forest_load_car, read_genesis_header};
-use forest_ipld::{recurse_links_hash, CidHashSet, DEFAULT_RECENT_STATE_ROOTS};
+use forest_ipld::{recurse_links_hash, CidHashSet};
 use forest_rpc_api::chain_api::ChainExportParams;
 use forest_rpc_client::chain_ops::*;
 use forest_utils::{io::parser::parse_duration, net::get_fetch_progress_from_file, retry};
@@ -21,7 +22,6 @@ use fvm_shared::clock::ChainEpoch;
 use log::info;
 use strfmt::strfmt;
 use tempfile::TempDir;
-use time::OffsetDateTime;
 use tokio::time::sleep;
 
 use super::*;
@@ -30,22 +30,29 @@ use crate::cli::{cli_error_and_die, handle_rpc_err};
 pub(crate) const OUTPUT_PATH_DEFAULT_FORMAT: &str =
     "forest_snapshot_{chain}_{year}-{month}-{day}_height_{height}.car";
 
+pub(crate) const OUTPUT_PATH_DEFAULT_COMPRESSED_FORMAT: &str =
+    "forest_snapshot_{chain}_{year}-{month}-{day}_height_{height}.car.zst";
+
 #[derive(Debug, Subcommand)]
 pub enum SnapshotCommands {
     /// Export a snapshot of the chain to `<output_path>`
     Export {
         /// Snapshot output path. Default to
-        /// `forest_snapshot_{chain}_{year}-{month}-{day}_height_{height}.car`
-        /// Date is in ISO 8601 date format.
+        /// `forest_snapshot_{chain}_{year}-{month}-{day}_height_{height}.car(.
+        /// zst)` Date is in ISO 8601 date format.
         /// Arguments:
         ///  - chain - chain name e.g. `mainnet`
         ///  - year
         ///  - month
         ///  - day
         ///  - height - the epoch
-        #[arg(short, default_value = OUTPUT_PATH_DEFAULT_FORMAT, verbatim_doc_comment)]
+        #[arg(short, default_value = ".", verbatim_doc_comment)]
         output_path: PathBuf,
-        /// Skip creating the checksum file.
+        /// Export in zstd compressed format
+        #[arg(long)]
+        compressed: bool,
+        /// Skip creating the checksum file. Only valid when `--compressed` is
+        /// not supplied.
         #[arg(long)]
         skip_checksum: bool,
         /// Skip writing to the snapshot `.car` file specified by
@@ -148,6 +155,7 @@ impl SnapshotCommands {
         match self {
             Self::Export {
                 output_path,
+                compressed,
                 skip_checksum,
                 dry_run,
             } => {
@@ -158,7 +166,7 @@ impl SnapshotCommands {
 
                 let epoch = chain_head.epoch();
 
-                let now = OffsetDateTime::now_utc();
+                let now = Utc::now();
 
                 let month_string = format!("{:02}", now.month() as u8);
                 let year = now.year();
@@ -177,7 +185,11 @@ impl SnapshotCommands {
                 ]);
 
                 let output_path = if output_path.is_dir() {
-                    output_path.join(OUTPUT_PATH_DEFAULT_FORMAT)
+                    output_path.join(if *compressed {
+                        OUTPUT_PATH_DEFAULT_COMPRESSED_FORMAT
+                    } else {
+                        OUTPUT_PATH_DEFAULT_FORMAT
+                    })
                 } else {
                     output_path.clone()
                 };
@@ -191,9 +203,10 @@ impl SnapshotCommands {
 
                 let params = ChainExportParams {
                     epoch,
-                    recent_roots: DEFAULT_RECENT_STATE_ROOTS,
+                    recent_roots: config.chain.recent_state_roots,
                     output_path,
                     tipset_keys: TipsetKeysJson(chain_head.key().clone()),
+                    compressed: *compressed,
                     skip_checksum: *skip_checksum,
                     dry_run: *dry_run,
                 };